Extract countries into its own file
